Edit: finished GET-script

+++ Slackbuilds/min/build/min.get-SB.sh
@@ -1,24 +1,47 @@
 #!/bin/env sh
-#### GET-CONFIG ####
+#### CONFIG ####
 SB_PKG="min"
 SB_VER="1.4.1"
 SB_EXT="tar.gz"
 SB_SOURCE="https://github.com/minbrowser/${SB_PKG}/archive/v${SB_VER}.${SB_EXT}"
 SB_REPO="https://notabug.org/jayvii/Slackbuilds"
-SB_DIR="$1"
-#### GET-CONFIG ####
+SB_DIR="/tmp/${SB_PKG}-SB"
+SB_BUILD="$1"
+#### CONFIG ####
+
+echo "Slackbuild for ${SB_PKG}, version ${SB_VER}"
+echo "=========="
+echo ""
 
 #### DOWNLOAD ####
-if [ -z "$SB_DIR" ]; then
-	echo "no target directory set."
-	echo "\"./${SB_PKG}.get-SB.sh /SOME/TARGET/FOLDER\""
-	echo "using \"/tmp/${SB_PKG}-SB/\" instead."
+if [ ${SB_BUILD} = "build-existing" ]; then
+	echo "User claims that Slackbuild and Source already exist in"
+	echo "\"/tmp/${SB_PKG}-SB/\""
+	echo "Skipping Download!"
+else
+	echo "Downloading..."
+	echo "using this Slackbuild-directory: \"/tmp/${SB_PKG}-SB/\"."
 	pause 2
-	SB_DIR="/tmp/${SB_PKG}-SB"
+	mkdir -p "$SB_DIR"
+	wget -nv "$SB_REPO/raw/master/Slackbuilds/${SB_PKG}/build/${SB_PKG}.SlackBuild" -O "${SB_DIR}/${SB_PKG}.SlackBuild"
+	wget -nv "$SB_REPO/raw/master/Slackbuilds/${SB_PKG}/build/${SB_PKG}.desc" -O "${SB_DIR}/${SB_PKG}.desc"
+	wget -nv "$SB_SOURCE" -O "${SB_DIR}/${SB_PKG}-source.${SB_EXT}"
+	ls -l "$SB_DIR"
 fi
-mkdir -p "$SB_DIR"
-wget -nv "$SB_SOURCE" -O "${SB_DIR}/${SB_PKG}-source.${SB_EXT}"
-wget -nv "$SB_REPO/raw/master/Slackbuilds/${SB_PKG}/build/${SB_PKG}.SlackBuild" -O "${SB_DIR}/${SB_PKG}.SlackBuild"
-wget -nv "$SB_REPO/raw/master/Slackbuilds/${SB_PKG}/build/${SB_PKG}.desc" -O "${SB_DIR}/${SB_PKG}.desc"
-ls -l "$SB_DIR"
 #### DOWNLOAD ####
+
+#### BUILD ####
+if [ ${SB_BUILD} = "build" ] || [ ${SB_BUILD} = "build-existing" ]; then
+	echo "Executing SlackBuild..."
+	cd ${SB_DIR}
+	sh -c ${SB_PKG}.SlackBuild
+else
+	echo "Build-parameter not set, so not building."
+	echo "If you whish to re-download the source and slackbuild, use the \"build\" parameter:"
+	echo "\"./${SB_PKG}.get-SB.sh build\""
+	echo "If you wish to use the already downloaded source and slackbuild, use the \"build-existing\" parameter:"
+	echo "\"./${SB_PKG}.get-SB.sh build-existing\""
+	echo "Or simply execute the already downloaded SlackBuild yourself:"
+	echo "\"./${SB_PKG}.SlackBuild\""
+fi
+#### BUILD ####
